fix: fixed handling of errors, logic of controllers, expected return types for routes and queries in the DAO

+++ code/server/src/controllers/productController.ts
@@ -1,17 +1,25 @@
+import dayjs from "dayjs";
 import { Product } from "../components/product";
 import ProductDAO from "../dao/productDAO";
-import { ProductNotFoundError, LowProductStockError, EmptyProductStockError } from "../errors/productError";
-
+import {
+    ProductNotFoundError,
+    LowProductStockError,
+    EmptyProductStockError,
+    ProductAlreadyExistsError,
+    IncorrectGroupingError,
+    NullCategoryGroupingError,
+    NullModelGroupingError,
+} from "../errors/productError";
 
 /**
  * Represents a controller for managing products.
  * All methods of this class must interact with the corresponding DAO class to retrieve or store data.
  */
 class ProductController {
-    private dao: ProductDAO
+    private dao: ProductDAO;
 
     constructor() {
-        this.dao = new ProductDAO
+        this.dao = new ProductDAO();
     }
 
     /**
@@ -24,11 +32,30 @@ class ProductController {
      * @param arrivalDate The optional date in which the product arrived.
      * @returns A Promise that resolves to nothing.
      */
-    async registerProducts(model: string, category: string, quantity: number, details: string | null, sellingPrice: number, arrivalDate: string | null): Promise<boolean> {
-        //const product: Product = { model, category, quantity, details, sellingPrice, arrivalDate };
-        return this.dao.addProduct(model, category, quantity, details, sellingPrice, arrivalDate);
-    }
+    async registerProducts(
+        model: string,
+        category: string,
+        quantity: number,
+        details: string | null,
+        sellingPrice: number,
+        arrivalDate: string | null,
+    ): Promise<boolean> {
+        if (await this.dao.getProductByModel(model)) {
+            throw new ProductAlreadyExistsError();
+        }
+        if (dayjs(arrivalDate).isAfter(dayjs())) {
+            throw new Error("Arrival date cannot be in the future");
+        }
 
+        return this.dao.registerProducts(
+            model,
+            category,
+            quantity,
+            details,
+            sellingPrice,
+            arrivalDate,
+        );
+    }
 
     /**
      * Increases the available quantity of a product through the addition of new units.
@@ -37,13 +64,26 @@ class ProductController {
      * @param changeDate The optional date in which the change occurred.
      * @returns A Promise that resolves to the new available quantity of the product.
      */
-    async changeProductQuantity(model: string, newQuantity: number, changeDate: string | null): Promise<number> {
+    async changeProductQuantity(
+        model: string,
+        newQuantity: number,
+        changeDate: string | null,
+    ): Promise<number> {
         const product = await this.dao.getProductByModel(model);
-        if (!product) throw new Error("Product not found");
 
-        product.quantity += newQuantity;
-        await this.dao.updateProduct(product);
-        return product.quantity; 
+        if (!product) throw new ProductNotFoundError();
+        if (dayjs(changeDate).isAfter(dayjs())) {
+            throw new Error("Change date cannot be in the future");
+        }
+        if (dayjs(changeDate).isBefore(dayjs(product.arrivalDate))) {
+            throw new Error("Change date cannot be before arrival date");
+        }
+
+        return this.dao.changeProductQuantity(
+            model,
+            product.quantity + newQuantity,
+            changeDate,
+        );
     }
 
     /**
@@ -53,17 +93,29 @@ class ProductController {
      * @param sellingDate The optional date in which the sale occurred.
      * @returns A Promise that resolves to the new available quantity of the product.
      */
-    async sellProduct(model: string, quantity: number, sellingDate: string | null): Promise<number> {
+    async sellProduct(
+        model: string,
+        quantity: number,
+        sellingDate: string | null,
+    ): Promise<number> {
         const product = await this.dao.getProductByModel(model);
-        if (!product) throw new Error("Product not found");
 
-        if (product.quantity < quantity) throw new Error("Insufficient quantity");
+        if (!product) throw new ProductNotFoundError();
+        if (dayjs(sellingDate).isAfter(dayjs())) {
+            throw new Error("Selling date cannot be in the future");
+        }
+        if (dayjs(sellingDate).isBefore(dayjs(product.arrivalDate))) {
+            throw new Error("Selling date cannot be before arrival date");
+        }
+        if (product.quantity === 0) throw new EmptyProductStockError();
+        if (product.quantity < quantity) throw new LowProductStockError();
 
-        product.quantity -= quantity;
-        await this.dao.updateProduct(product);
-        return product.quantity;
+        return this.dao.sellProduct(
+            model,
+            product.quantity - quantity,
+            sellingDate,
+        );
     }
-
 
     /**
      * Returns all products in the database, with the option to filter them by category or model.
@@ -72,13 +124,24 @@ class ProductController {
      * @param model An optional parameter. It can only be present if grouping is equal to "model" (in which case it must be present and not empty).
      * @returns A Promise that resolves to an array of Product objects.
      */
-    async getProducts(grouping: string | null, category: string | null, model: string | null): Promise<Product[]> {
-        if (grouping === "category" && category) {
-            return this.dao.getProductsByCategory(category);
-        } else if (grouping === "model" && model) {
-            return this.dao.getProductsByModel(model);
-        } else {
-            return this.dao.getAllProducts();
+    async getProducts(
+        grouping: string | null,
+        category: string | null,
+        model: string | null,
+    ): Promise<Product[]> {
+        switch (grouping) {
+            case "category":
+                if (!category) throw new NullCategoryGroupingError();
+                return this.dao.getProductsByCategory(category);
+            case "model":
+                if (!model) throw new NullModelGroupingError();
+                if (!(await this.dao.getProductByModel(model))) {
+                    throw new ProductNotFoundError();
+                }
+                return this.dao.getProductsByModel(model);
+            default:
+                if (category || model) throw new IncorrectGroupingError();
+                return this.dao.getAllProducts();
         }
     }
 
@@ -89,15 +152,27 @@ class ProductController {
      * @param model An optional parameter. It can only be present if grouping is equal to "model" (in which case it must be present and not empty).
      * @returns A Promise that resolves to an array of Product objects.
      */
-    async getAvailableProducts(grouping: string | null, category: string | null, model: string | null): Promise<Product[]> {
-        if (grouping === "category" && category) {
-            return this.dao.getAvailableProductsByCategory(category);
-        } else if (grouping === "model" && model) {
-            return this.dao.getAvailableProductsByModel(model);
-        } else {
-            return this.dao.getAllAvailableProducts();
+    async getAvailableProducts(
+        grouping: string | null,
+        category: string | null,
+        model: string | null,
+    ): Promise<Product[]> {
+        switch (grouping) {
+            case "category":
+                if (!category) throw new NullCategoryGroupingError();
+                return this.dao.getAvailableProductsByCategory(category);
+            case "model":
+                if (!model) throw new NullModelGroupingError();
+                if (!(await this.dao.getProductByModel(model))) {
+                    throw new ProductNotFoundError();
+                }
+                return this.dao.getAvailableProductsByModel(model);
+            default:
+                if (category || model) throw new IncorrectGroupingError();
+                return this.dao.getAllAvailableProducts();
         }
     }
+
     /**
      * Deletes all products.
      * @returns A Promise that resolves to `true` if all products have been successfully deleted.
@@ -106,17 +181,18 @@ class ProductController {
         return this.dao.deleteAllProducts();
     }
 
-
     /**
      * Deletes one product, identified by its model
      * @param model The model of the product to delete
      * @returns A Promise that resolves to `true` if the product has been successfully deleted.
      */
     async deleteProduct(model: string): Promise<boolean> {
+        if (!(await this.dao.getProductByModel(model))) {
+            throw new ProductNotFoundError();
+        }
+
         return this.dao.deleteProduct(model);
     }
-
-
 }
 
 export default ProductController;
